Update transient dynamics
Update the Notebook to provide an example/default values to the notebook, allowing the user to see the initial functionality of each block of code. Update the terminology used in the sections of code to make it more comprehensive and change the save names for the files to be more descriptive.